REFACTOR project directory structure
