+++ goat/project/configuration/project_configuration_variant.py
