+++ goat/project/configuration/raw/build_configuration_variant.py
