+++ goat/project/configuration/raw/packages_configuration.py
